% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedData-class.R, R/dimRedResult-class.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,dimRedData-method}
\alias{as.data.frame,dimRedResult-method}
\title{Convert to \code{data.frame}}
\usage{
as.data.frame(x, row.names, optional, ...)

\S4method{as.data.frame}{dimRedData}(x, meta.prefix = "meta.",
  data.prefix = "")

\S4method{as.data.frame}{dimRedResult}(x, org.data.prefix = "org.",
  meta.prefix = "meta.", data.prefix = "")
}
\arguments{
\item{x}{dimRedResult/dimRedData object.}

\item{row.names}{unused}

\item{optional}{unused}

\item{...}{unused}

\item{meta.prefix}{prefix for the meta data variables.}

\item{data.prefix}{for dimRedResult objects: prefix for embedded
dimensions, for dimRedData objects: prefix for the variable
names.}

\item{org.data.prefix}{for dimRedResult objects: prefix for the
original variables.}
}
\description{
Convert a dimRedData or dimRedResult object to a data.frame.
}
\details{
To avoid column name collisions in the resulting \code{data.frame}
prefixes can be assigned. The parameters \code{row.names},
\code{optional}, and \code{...} are not used.
}
\examples{
as.data.frame(loadDataSet("Iris"), meta.prefix = "")
as.data.frame(embed(loadDataSet("Iris"), "pca"), org.data.prefix = "", meta.prefix = "")

}
\seealso{
Other dimRedData: \code{\link{as.dimRedData}},
  \code{\link{dimRedData}}

Other dimRedResult: \code{\link{dimRedResult-class}}
}

