% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedData-class.R
\docType{methods}
\name{as.dimRedData}
\alias{as.dimRedData}
\alias{as.dimRedData,formula-method}
\title{Convert to \code{dimRedData}}
\usage{
as.dimRedData(x, ...)

\S4method{as.dimRedData}{formula}(x, data)
}
\arguments{
\item{x}{the formula, left hand side is assigned to meta slot
right hand side is assigned to data slot.}

\item{...}{parameters}

\item{data}{a data frame}
}
\description{
Convert a \code{data.frame} to a dimRedData object using a formula
}
\seealso{
Other dimRedData: \code{\link{as.data.frame}},
  \code{\link{dimRedData}}
}

