% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\docType{methods}
\name{AUC_R_NX,dimRedResult-method}
\alias{AUC_R_NX,dimRedResult-method}
\alias{AUC_lnK}
\title{Method AUC_R_NX}
\usage{
\S4method{AUC_R_NX}{dimRedResult}(object)
}
\arguments{
\item{object}{of class dimRedResult}
}
\description{
Calculate the Area under the R_NX(K) curve for logarithmic K,
used in Lee et. al. (2013).
}
\seealso{
Other Quality scores for dimensionality reduction: \code{\link{LCMC,dimRedResult-method}},
  \code{\link{Q_NX,dimRedResult-method}},
  \code{\link{Q_global,dimRedResult-method}},
  \code{\link{Q_local,dimRedResult-method}},
  \code{\link{R_NX,dimRedResult-method}},
  \code{\link{cophenetic_correlation,dimRedResult-method}},
  \code{\link{distance_correlation,dimRedResult-method}},
  \code{\link{mean_R_NX,dimRedResult-method}},
  \code{\link{quality,dimRedResult-method}},
  \code{\link{reconstruction_rmse,dimRedResult-method}},
  \code{\link{total_correlation,dimRedResult-method}}
}

