% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\docType{methods}
\name{cophenetic_correlation,dimRedResult-method}
\alias{cophenetic_correlation}
\alias{cophenetic_correlation,dimRedResult-method}
\title{Method cophenetic_correlation}
\usage{
\S4method{cophenetic_correlation}{dimRedResult}(object, d = stats::dist,
  cor_method = "pearson")
}
\arguments{
\item{object}{of class dimRedResult}

\item{d}{the distance function to use.}

\item{cor_method}{The correlation method.}
}
\description{
Calculate the correlation between the distance matrices in high and
low dimensioal space.
}
\seealso{
Other Quality scores for dimensionality reduction: \code{\link{AUC_lnK_R_NX,dimRedResult-method}},
  \code{\link{LCMC,dimRedResult-method}},
  \code{\link{Q_NX,dimRedResult-method}},
  \code{\link{Q_global,dimRedResult-method}},
  \code{\link{Q_local,dimRedResult-method}},
  \code{\link{R_NX,dimRedResult-method}},
  \code{\link{distance_correlation,dimRedResult-method}},
  \code{\link{mean_R_NX,dimRedResult-method}},
  \code{\link{quality,dimRedResult-method}},
  \code{\link{reconstruction_rmse,dimRedResult-method}},
  \code{\link{total_correlation,dimRedResult-method}}
}

