% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmap.R
\docType{data}
\name{diffmap}
\alias{diffmap}
\title{Diffusion Maps}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
diffmap
}
\description{
Instance of \code{\link{dimRedMethod}} to construct a diffuson maps.
}
\details{
For details see \code{\link[diffusionMap]{diffuse}}.
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- diffmap@fun(dat, diffmap@stdpars)

plot(emb@data@data)

}
\keyword{datasets}

