% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed-class.R
\docType{class}
\name{dimRed-class}
\alias{dimRed-class}
\title{S4 Class to hold data, method and parameters for dimensionality reduction}
\description{
S4 Class to hold data, method and parameters for dimensionality reduction
}
\section{Slots}{

\describe{
\item{\code{data}}{holds the original data}

\item{\code{method}}{holds the method}

\item{\code{the}}{parameters used with method}
}}

