% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\docType{package}
\name{dimRed-package}
\alias{dimRed}
\alias{dimRed-package}
\title{A package to aggregate dimensionality reduction techniques.}
\description{
dimRed collects dimensionality reduction methods that are implemented in R and give them a common interface.
}
\details{
TODO: add a nice description
}

