% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedData-class.R
\docType{class}
\name{dimRedData-class}
\alias{dimRedData-class}
\title{An S4 class to hold data for dimensionality reduction}
\description{
An S4 class to hold data for dimensionality reduction
}
\section{Slots}{

\describe{
\item{\code{data}}{of class \code{matrix}, holds the data, observations in
rows, variables in columns}
}}

