% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedMethod-class.R
\docType{class}
\name{dimRedMethod-class}
\alias{dimRedMethod}
\alias{dimRedMethod-class}
\alias{dimRedMethodList}
\title{An S4 class that makes sure that the wrappers for dimensionality
reduction work.}
\usage{
dimRedMethodList()
}
\description{
An S4 class that makes sure that the wrappers for dimensionality
reduction work.
}
\section{Slots}{

\describe{
\item{\code{fun}}{function call}
}}

