% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedMethod-class.R
\name{dimRedMethodList}
\alias{dimRedMethodList}
\title{dimRedMethodList}
\usage{
dimRedMethodList(filter = FALSE)
}
\arguments{
\item{filter}{filter methods by methods that have their dependencies installed}
}
\value{
a character vector with the names of classes that inherit
    from \code{dimRedMethod}.
}
\description{
Get the names of all methods for dimensionality reduction.
}
\details{
Returns the name of all classes that inherit from
\code{\link{dimRedMethod-class}} to use with \code{\link{embed}}.
}
\examples{
dimRedMethodList()

}
\seealso{
Other dimensionality reduction methods: 
\code{\link{DRR-class}},
\code{\link{DiffusionMaps-class}},
\code{\link{DrL-class}},
\code{\link{FastICA-class}},
\code{\link{FruchtermanReingold-class}},
\code{\link{HLLE-class}},
\code{\link{Isomap-class}},
\code{\link{KamadaKawai-class}},
\code{\link{MDS-class}},
\code{\link{NNMF-class}},
\code{\link{PCA-class}},
\code{\link{PCA_L1-class}},
\code{\link{UMAP-class}},
\code{\link{dimRedMethod-class}},
\code{\link{kPCA-class}},
\code{\link{nMDS-class}},
\code{\link{tSNE-class}}
}
\concept{dimensionality reduction methods}
