% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedMethodPars-class.R
\docType{class}
\name{dimRedMethodPars-class}
\alias{dimRedMethodPars-class}
\title{S4 Class to store method parameters}
\description{
at the moment nothing more than a wrapper around a list
}
\section{Slots}{

\describe{
\item{\code{pars}}{a named list of parameters}
}}

