% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embed.R
\docType{class}
\name{DrL-class}
\alias{DrL}
\alias{DrL-class}
\title{Distributed Recursive Graph Layout}
\description{
Instance of \code{\link{dimRedMethod}} for the Distributed Recursive Graph Layout algorithm.
}
\details{
For details see \code{\link[igraph]{layout_with_drl}}
}
\examples{

dat <- loadDataSet("Swiss Roll")
drl <- DrL()
drgl <- drl@fun(dat, drl@stdpars)

plot(drgl@data@data)


}

