% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embed.R
\docType{data}
\name{drl}
\alias{drl}
\title{Distributed Recursive Graph Layout}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
drl
}
\description{
Instance of \code{\link{dimRedMethod}} for the Distributed Recursive Graph Layout algorithm.
}
\details{
For details see \code{\link[igraph]{layout_with_drl}}
}
\examples{

dat <- loadDataSet("Swiss Roll")
drgl <- drl@fun(dat, drl@stdpars)

plot(drgl@data@data)


}
\keyword{datasets}

