% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drr.R
\docType{data}
\name{drr}
\alias{drr}
\title{implements Dimensionality Rediction via Regression for the use with dimRed}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
drr
}
\arguments{
\item{data}{an object of class \code{\link{dimRedData-class}}}

\item{pars}{an object of class \code{\link{dimRedMethodPars-class}}}
}
\value{
an object of class \code{\link{dimRedResult-class}}
}
\description{
uses \code{\link[DRR]{DRR}} internally. Has methods for forward and inverse calculations.
}
\examples{

dat <- loadDataSet("3D S Curve", n = 500)
emb <- drr@fun(dat, pars = list(ndim = 2)), file = "NUL", type = 'output'

plot(emb@data@data)


}
\keyword{datasets}

