% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{dispatches the different methods for dimensionality reduction}
\usage{
embed(data, method, keep.org.data = TRUE, ...)
}
\arguments{
\item{data}{object of class \code{dimRedData}}

\item{method}{character vector naming one of the dimensionality
reduction techniques}

\item{...}{the pameters, internally passed as a list to the
dimensionality reduction method as \code{pars = list(...)}}
}
\value{
an object of class \code{dimRedResult}
}
\description{
wraps around all dimensionality reduction functions.
}
\details{
method must be one of \code{"\link{graph_kk}",
"\link{graph_drl}", "\link{graph_fr}",
"\link{isomap}", "\link{tsne}", "\link{nmds}",
"\link{mds}", "\link{ica}", "\link{pca}",
"\link{lle}", "\link{loe}", "\link{soe}",
"\link{leim}", "\link{kpca}"}
}
\examples{
embed_methods <- c("graph_kk", "graph_drl", "graph_fr", "isomap", "tsne", "nmds",
                   "mds", "ica", "pca", "lle", "drr",
                   #"loe", "soe",
                   "leim", "kpca")
quality_methods <- c("Q_local", "Q_global", "mean_R_NX", "total_correlation",
                     "cophenetic_correlation", "distance_correlation",
                     "reconstruction_rmse")
scurve <- loadDataSet("3D S Curve", n = 500)

quality_results <- matrix(NA, length(embed_methods), length(quality_methods),
                              dimnames = list(embed_methods, quality_methods))
embedded_data <- list()

for(e in embed_methods) {
  message("embedding: ", e)
  embedded_data[[e]] <- embed(scurve, e)
  for(q in quality_methods) {
    message("  quality: ", q)
    try(quality_results[e,q] <- quality(embedded_data[[e]], q))
  }
}

print(quality_results)



}

