% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastica.R
\docType{class}
\name{FastICA-class}
\alias{FastICA}
\alias{FastICA-class}
\title{FastICA}
\description{
Instance of \code{\link{dimRedMethod}} for the fast ICA algoritm.
}
\details{
For details see \code{\link[fastICA]{fastICA}}.
}
\examples{
dat <- loadDataSet("3D S Curve")
fastica <- FastICA()
emb <- fastica@fun(dat, pars = list(ndim = 2))

plot(emb@data@data)


}

