% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastica.R
\docType{data}
\name{fastica}
\alias{fastica}
\title{FastICA}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
fastica
}
\description{
the fast ICA algoritm
}
\details{
for details see \code{\link[fastICA]{fastICA}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- fastica@fun(dat, pars = list(ndim = 2))

plot(emb@data@data)


}
\keyword{datasets}

