% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed-class.R
\name{fit}
\alias{fit}
\title{Fit a dimred object.}
\usage{
fit(object)
}
\arguments{
\item{x}{the dimRed object}
}
\description{
Fits a dimred object.
}
\details{
the same can be achieved by e.g. \code{fastica@fun(dat, pars = ...)}
}

