% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embed.R
\docType{data}
\name{fruchterman_reingold}
\alias{fruchterman_reingold}
\title{Fruchterman Reingold Graph Layout}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
fruchterman_reingold
}
\description{
Instance of \code{\link{dimRedMethod}} for the Fruchterman Reingold Graph Layout algorithm.
}
\details{
For details see \code{\link[igraph]{layout_with_fr}}
}
\examples{

dat <- loadDataSet("Swiss Roll")
fr <- fruchterman_reingold@fun(dat, fruchterman_reingold@stdpars)

plot(fr@data@data)


}
\keyword{datasets}

