% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embed.R
\docType{class}
\name{FruchtermanReingold-class}
\alias{FruchtermanReingold}
\alias{FruchtermanReingold-class}
\title{Fruchterman Reingold Graph Layout}
\description{
Instance of \code{\link{dimRedMethod}} for the Fruchterman Reingold Graph Layout algorithm.
}
\details{
For details see \code{\link[igraph]{layout_with_fr}}
}
\examples{

dat <- loadDataSet("Swiss Roll")
fruchterman_reingold <- FruchtermanReingold()
fr <- fruchterman_reingold@fun(dat, fruchterman_reingold@stdpars)

plot(fr@data@data)


}

