% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getDimRedData}
\alias{getDimRedData}
\title{Method getDimRedData}
\usage{
getDimRedData(object, ...)
}
\arguments{
\item{object}{The object to extract data from.}

\item{...}{other arguments.}
}
\description{
Extract dimRedData.
}

