% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getNDim}
\alias{getNDim}
\title{Method getNDim}
\usage{
getNDim(object, ...)
}
\arguments{
\item{object}{The object to get the dimensions from.}

\item{...}{other arguments.}
}
\description{
Extract the number of embedding dimensions.
}
