% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getOrgData}
\alias{getOrgData}
\title{Method getOrgData}
\usage{
getOrgData(object, ...)
}
\arguments{
\item{object}{The object to extract data from.}

\item{...}{other arguments.}
}
\description{
Extract the Original data.
}

