% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getOtherData}
\alias{getOtherData}
\title{Method getOtherData}
\usage{
getOtherData(object, ...)
}
\arguments{
\item{object}{The object to extract data from.}

\item{...}{other arguments.}
}
\description{
Extract other data produced by a dimRedMethod
}
