% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomap.R
\docType{class}
\name{Isomap-class}
\alias{Isomap}
\alias{Isomap-class}
\title{Isomap embedding}
\description{
Instance of \code{\link{dimRedMethod}} for Isomap.
}
\details{
Create a k nearest neighbor graph, compute geodesic distance on the
graph, use classical scaling for the embedding.
The landmark version is implemented via \code{isomap@apply()}.

Parameters to set are
\itemize{
\item k. The number of neighbors
\item ndim. The number of dimensions
\item eps. if larger than 0, approximate the nearest neighbors.
}
}
\examples{
dat <- loadDataSet("3D S Curve")
isomap <- Isomap()
emb <- isomap@fun(dat, isomap@stdpars)

plot(emb@data@data)

}

