% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embed.R
\docType{data}
\name{kamada_kawai}
\alias{kamada_kawai}
\title{implements Graph Embedding methods via iGraph}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
kamada_kawai
}
\description{
implements Graph Embedding methods via iGraph
}
\examples{

dat <- loadDataSet("Swiss Roll")

em <- fruchterman_reingold@fun(dat)
em2 <- kamada_kawai@fun(dat)
em3 <- drl@fun(dat)

plot(em@data@data)
plot(em2@data@data)
plot(em3@data@data)


}
\keyword{datasets}

