% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embed.R
\docType{class}
\name{KamadaKawai-class}
\alias{KamadaKawai}
\alias{KamadaKawai-class}
\title{Graph Embedding via the Kamada Kawai Algorithm}
\description{
Instance of \code{\link{dimRedMethod}} for the Kamada Kawai Algorithm.
}
\details{
For details see \code{\link[igraph]{layout_with_kk}}.
}
\examples{
dat <- loadDataSet("Swiss Roll")
kamada_kawai <- KamadaKawai()
kk <- kamada_kawai@fun(dat, kamada_kawai@stdpars)

plot(kk@data@data)


}

