% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\docType{class}
\name{kPCA-class}
\alias{kPCA}
\alias{kPCA-class}
\title{Kernel PCA}
\description{
Instance of \code{\link{dimRedMethod}} for Kernel PCA.
}
\details{
For details, see \code{\link[kernlab]{kpca}}
}
\examples{

dat <- loadDataSet("3D S Curve")
kpca <- kPCA()
emb <- kpca@fun(dat, list(kernel = 'rbfdot',
                          kpar = list(sigma = 0.1),
                          ndim = 2))

plot(emb@data@data)

}

