% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\docType{data}
\name{kpca}
\alias{kpca}
\title{Kernel PCA}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
kpca
}
\description{
Instance of \code{\link{dimRedMethod}} for Kernel PCA.
}
\details{
For details, see \code{\link[kernlab]{kpca}}
}
\examples{

dat <- loadDataSet("3D S Curve")
emb <- kpca@fun(dat, list(kernel = 'rbfdot',
                          kpar = list(sigma = 0.1),
                          ndim = 2))

plot(emb@data@data)

}
\keyword{datasets}

