% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leim.R
\docType{class}
\name{LaplacianEigenmaps-class}
\alias{LaplacianEigenmaps}
\alias{LaplacianEigenmaps-class}
\title{Laplacian Eigenmaps}
\description{
Instance of \code{\link{dimRedMethod}} for Laplacian Eigenmaps.
}
\details{
For details see \code{\link[loe]{spec.emb}}
}
\examples{
dat <- loadDataSet("3D S Curve")
leim <- LaplacianEigenmaps()
emb <- leim@fun(dat, leim@stdpars)


plot(emb@data@data)


}

