% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leim.R
\docType{data}
\name{leim}
\alias{leim}
\title{laplacian eigenmaps}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
leim
}
\description{
Fit laplacian eigenmaps
}
\details{
for details see \code{\link[loe]{spec.emb}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- leim@fun(dat)


plot(emb@data@data)


}
\keyword{datasets}

