% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lle.R
\docType{class}
\name{LLE-class}
\alias{LLE}
\alias{LLE-class}
\title{Local Linear Embedding}
\description{
Instance of \code{\link{dimRedMethod}} for Local Linear Embedding.
}
\details{
For details see \code{\link[lle]{lle}}
}
\examples{
dat <- loadDataSet("3D S Curve")
lle <- LLE()
emb <- lle@fun(dat, lle@stdpars)


plot(emb@data@data)

}

