% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lle.R
\docType{data}
\name{lle}
\alias{lle}
\title{Local Linear Embedding}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
lle
}
\description{
Instance of \code{\link{dimRedMethod}} for Local Linear Embedding.
}
\details{
For details see \code{\link[lle]{lle}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- lle@fun(dat, lle@stdpars)


plot(emb@data@data)

}
\keyword{datasets}

