% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\name{loadDataSet}
\alias{loadDataSet}
\title{Example Data Sets for dimensionality reduction}
\usage{
loadDataSet(name = "Swiss Roll", n = 2000, sigma = 0.05)
}
\arguments{
\item{name}{a character vector that specifies the name of the data set}

\item{n}{in generated data sets the number of points to be generated}

\item{sigma}{standard deviation of the noise added}
}
\description{
a compilation of standard data sets that are often being used to
showcase dimensionality reduction techniques.
}
\details{
name should be one of
  \code{c('Swiss Roll', 'broken Swiss Roll',
          'Helix', 'Twin Peaks', 'Sphere', 'Ball',
          '3D S Curve', 'variable Noise Helix', 'Iris')}
}

