% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loe.R
\docType{class}
\name{LOE-class}
\alias{LOE}
\alias{LOE-class}
\title{Local Ordinal Embedding}
\description{
Instance of \code{\link{dimRedMethod}} for Local Ordinal Embedding.
}
\details{
For details see \code{\link[loe]{LOE}}
}
\examples{
# for whatever reason the loe package has problems if I run this
# with R CMD check, running it in the REPL works just fine
dat <- loadDataSet("Iris")
loe <- LOE()
emb <- loe@fun(dat, loe@stdpars)


plot(emb@data@data)

}

