% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomap.R
\name{makeKNNgraph}
\alias{makeKNNgraph}
\title{input data(matrix or data frame) return knn graph
implements "smart" choices on RANN::nn2 parameters
we ignore radius search
TODO: find out a good limit to switch from kd to bd trees}
\usage{
makeKNNgraph(x, k, eps = 0)
}
\description{
input data(matrix or data frame) return knn graph
implements "smart" choices on RANN::nn2 parameters
we ignore radius search
TODO: find out a good limit to switch from kd to bd trees
}

