% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\docType{class}
\name{MDS-class}
\alias{MDS}
\alias{MDS-class}
\title{Metric Dimensional Scaling}
\description{
Instance of \code{\link{dimRedMethod}} for Metric Dimensional Scaling.
}
\details{
For details see \code{\link[stats]{cmdscale}}
}
\examples{
dat <- loadDataSet("3D S Curve")
mds <- MDS()
emb <- mds@fun(dat, mds@stdpars)

plot(emb@data@data)

}

