% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\docType{data}
\name{mds}
\alias{mds}
\title{Metric Dimensional Scaling}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
mds
}
\description{
Instance of \code{\link{dimRedMethod}} for Metric Dimensional Scaling.
}
\details{
For details see \code{\link[stats]{cmdscale}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- mds@fun(dat, mds@stdpars)

plot(emb@data@data)

}
\keyword{datasets}

