% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ndims}
\alias{ndims}
\title{Method ndims}
\usage{
ndims(object, ...)
}
\arguments{
\item{object}{To extract the number of dimensions from.}

\item{...}{Arguments for further methods}
}
\description{
Extract the number of dimensions.
}

