% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\docType{class}
\name{nMDS-class}
\alias{nMDS}
\alias{nMDS-class}
\title{Non-Metric Dimensional Scaling}
\description{
Instance of \code{\link{dimRedMethod}} for the Non-Metric Dimensional Scaling.
}
\details{
For details see \code{\link[vegan]{monoMDS}}
}
\examples{
dat <- loadDataSet("3D S Curve")
nmds <- nMDS()
emb <- nmds@fun(dat, nmds@stdpars)

plot(emb@data@data)

}

