% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\docType{data}
\name{nmds}
\alias{nmds}
\title{Non-Metric Dimensional Scaling}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
nmds
}
\description{
Instance of \code{\link{dimRedMethod}} for the Non-Metric Dimensional Scaling.
}
\details{
For details see \code{\link[vegan]{monoMDS}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- nmds@fun(dat, nmds@stdpars)

plot(emb@data@data)

}
\keyword{datasets}

