% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\docType{class}
\name{PCA-class}
\alias{PCA}
\alias{PCA-class}
\title{Principal Component Analysis}
\description{
Instance of \code{\link{dimRedMethod}} for Principal Component Analysis.
}
\details{
For details see \code{\link{prcomp}}.
}
\examples{
dat <- loadDataSet("3D S Curve")
pca <- PCA()
emb <- pca@fun(dat, pca@stdpars)

plot(emb, type = "2vars")
plot(emb@inverse(emb@data), type = "3vars")

}

