% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\docType{data}
\name{pca}
\alias{pca}
\title{implements PCA for the use with dimRed}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
pca
}
\arguments{
\item{data}{an object of class \code{\link{dimRedData-class}}.}

\item{pars}{an object of class \code{\link{dimRedMethodPars-class}}}
}
\value{
an object of class \code{\link{dimRedResult-class}}


dat <- loadDataSet("3D S Curve")
emb <- pca@fun(dat)

plot(emb@data@data)
}
\description{
uses the \code{\link{prcomp}} function internally. Has methods for
forward and inverse calculations.
}
\keyword{datasets}

