% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pdist2}
\alias{pdist2}
\title{Squared euclidean distance between points in A and B
taken from http://blog.felixriedel.com/2013/05/pairwise-distances-in-r/}
\usage{
pdist2(A, B)
}
\description{
Squared euclidean distance between points in A and B
taken from http://blog.felixriedel.com/2013/05/pairwise-distances-in-r/
}

