% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,dimRedData,ANY-method}
\alias{plot,dimRedData,ANY-method}
\title{Plotting of dimRed* objects}
\usage{
\S4method{plot}{dimRedData,ANY}(x, y, ...)
}
\arguments{
\item{x}{dimRedResult/dimRedData class, e.g. output of
embedded/loadDataSet}

\item{type}{plot type, one of \code{c("pairs", "parallel", "2d",
"3d")}}

\item{colors}{the columns of the meta slot to use for coloring}

\item{vars}{the axes of the embedding to use for plotting}
}
\description{
Plots a object of class dimRedResult and dimRedData
}
\details{
somewhat inflexible, if you require custom colors or want to modify
graphical parameters, write your own plotting function.
}
\examples{
scurve = loadDataSet("3D S Curve")
plot(scurve, "pairs")

}

