% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,dimRedData,missing-method}
\alias{plot,dimRedData,missing-method}
\title{Plotting of dimRed* objects}
\usage{
\S4method{plot}{dimRedData,missing}(x, y = NULL, type = "pairs",
  vars = seq_len(ncol(x@data)), col = seq_len(min(3, ncol(x@meta))), ...)
}
\arguments{
\item{x}{dimRedResult/dimRedData class, e.g. output of
embedded/loadDataSet}

\item{y}{Ignored}

\item{type}{plot type, one of \code{c("pairs", "parallel", "2d",
"3d")}}

\item{vars}{the axes of the embedding to use for plotting}

\item{col}{the columns of the meta slot to use for coloring, can be
referenced as the column names or number of x@data}

\item{...}{handed over to the underlying plotting function.}
}
\description{
Plots a object of class dimRedResult and dimRedData
}
\details{
somewhat inflexible, if you require custom colors or want to modify
graphical parameters, write your own plotting function.
}
\examples{
scurve = loadDataSet("3D S Curve")
plot(scurve, type = "pairs", main = "pairs plot of S curve")
plot(scurve, type = "parpl")
plot(scurve, type = "2vars", vars = c("y", "z"))

}

