% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,dimRedData,ANY-method}
\alias{plot,dimRedResult,ANY-method}
\title{Plotting of dimRed* objects}
\usage{
plot(x, y, ...)

\S4method{plot}{dimRedData,ANY}(x, type = "pairs",
  vars = seq_len(ncol(x@data)), col = seq_len(min(3, ncol(x@meta))), ...)

\S4method{plot}{dimRedResult,ANY}(x, type = "pairs",
  vars = seq_len(ncol(x@data@data)), col = seq_len(min(3,
  ncol(x@data@meta))), ...)
}
\arguments{
\item{x}{dimRedResult/dimRedData class, e.g. output of
embedded/loadDataSet}

\item{y}{Ignored}

\item{...}{handed over to the underlying plotting function.}

\item{type}{plot type, one of \code{c("pairs", "parallel", "2d",
"3d")}}

\item{vars}{the axes of the embedding to use for plotting}

\item{col}{the columns of the meta slot to use for coloring, can be
referenced as the column names or number of x@data}
}
\description{
Plots a object of class dimRedResult and dimRedData
}
\details{
somewhat inflexible, if you require custom colors or want to modify
graphical parameters, write your own plotting function.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = dimRedData,y = ANY}: Plot dimRedData objects

\item \code{x = dimRedResult,y = ANY}: plot dimRedResult objects
}}
\examples{
scurve = loadDataSet("3D S Curve")
plot(scurve, type = "pairs", main = "pairs plot of S curve")
plot(scurve, type = "parpl")
plot(scurve, type = "2vars", vars = c("y", "z"))
plot(scurve, type = "3vars")

}

