% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print}
\alias{print}
\title{Method print}
\usage{
print(x, ...)
}
\arguments{
\item{x}{The object to be printed.}

\item{...}{Other arguments for printing.}
}
\description{
Imports the print method into the package namespace.
}
