% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\docType{methods}
\name{Q_global,dimRedResult-method}
\alias{Q_global}
\alias{Q_global,dimRedResult-method}
\title{Method Q_global}
\usage{
\S4method{Q_global}{dimRedResult}(object)
}
\arguments{
\item{object}{of class dimRedResult}
}
\description{
Calculate the Q_global score to assess the quality of a dimensionality reduction.
}
\seealso{
Other Quality scores for dimensionality reduction: \code{\link{AUC_R_NX,dimRedResult-method}},
  \code{\link{LCMC,dimRedResult-method}},
  \code{\link{Q_NX,dimRedResult-method}},
  \code{\link{Q_local,dimRedResult-method}},
  \code{\link{R_NX,dimRedResult-method}},
  \code{\link{cophenetic_correlation,dimRedResult-method}},
  \code{\link{distance_correlation,dimRedResult-method}},
  \code{\link{mean_R_NX,dimRedResult-method}},
  \code{\link{quality,dimRedResult-method}},
  \code{\link{reconstruction_rmse,dimRedResult-method}},
  \code{\link{total_correlation,dimRedResult-method}}
}

