% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\docType{methods}
\name{R_NX,dimRedResult-method}
\alias{R_NX,dimRedResult-method}
\alias{R_NX}
\title{Method R_NX}
\usage{
\S4method{R_NX}{dimRedResult}(object, ndim = getNDim(object))
}
\arguments{
\item{object}{of class dimRedResult}

\item{ndim}{the number of dimensions to take from the embedded data.}
}
\description{
Calculate the R_NX score from Lee et. al. (2013) which shows the neighborhood
preservation for the Kth nearest neighbors, corrected for random point
distributions and scaled to range [0, 1].
}
\seealso{
Other Quality scores for dimensionality reduction: \code{\link{AUC_lnK_R_NX,dimRedResult-method}},
  \code{\link{LCMC,dimRedResult-method}},
  \code{\link{Q_NX,dimRedResult-method}},
  \code{\link{Q_global,dimRedResult-method}},
  \code{\link{Q_local,dimRedResult-method}},
  \code{\link{cophenetic_correlation,dimRedResult-method}},
  \code{\link{distance_correlation,dimRedResult-method}},
  \code{\link{mean_R_NX,dimRedResult-method}},
  \code{\link{plot_R_NX}},
  \code{\link{quality,dimRedResult-method}},
  \code{\link{reconstruction_error,dimRedResult-method}},
  \code{\link{reconstruction_rmse,dimRedResult-method}},
  \code{\link{total_correlation,dimRedResult-method}}
}
