% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\docType{methods}
\name{reconstruction_rmse,dimRedResult-method}
\alias{reconstruction_rmse}
\alias{reconstruction_rmse,dimRedResult-method}
\title{Method reconstruction_rmse}
\usage{
\S4method{reconstruction_rmse}{dimRedResult}(object)
}
\arguments{
\item{object}{of class dimRedResult}
}
\description{
Calculate the reconstruction root mean squared error a dimensionality reduction, the method must have an inverse mapping.
}
\seealso{
Other Quality scores for dimensionality reduction: \code{\link{AUC_lnK_R_NX,dimRedResult-method}},
  \code{\link{LCMC,dimRedResult-method}},
  \code{\link{Q_NX,dimRedResult-method}},
  \code{\link{Q_global,dimRedResult-method}},
  \code{\link{Q_local,dimRedResult-method}},
  \code{\link{R_NX,dimRedResult-method}},
  \code{\link{cophenetic_correlation,dimRedResult-method}},
  \code{\link{distance_correlation,dimRedResult-method}},
  \code{\link{mean_R_NX,dimRedResult-method}},
  \code{\link{quality,dimRedResult-method}},
  \code{\link{total_correlation,dimRedResult-method}}
}

