% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soe.R
\docType{class}
\name{SOE-class}
\alias{SOE}
\alias{SOE-class}
\title{Soft Ordinal Embedding}
\description{
Instance of \code{\link{dimRedMethod}} for Soft Ordinal Embedding.
}
\details{
For details see \code{\link[loe]{SOE}}.
}
\examples{
dat <- loadDataSet("3D S Curve", n = 50)
soe <- SOE()
emb <- soe@fun(dat, soe@stdpars)


plot(emb@data@data)


}

