% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soe.R
\docType{data}
\name{soe}
\alias{soe}
\title{Soft Ordinal Embedding}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
soe
}
\description{
Instance of \code{\link{dimRedMethod}} for Soft Ordinal Embedding.
}
\details{
For details see \code{\link[loe]{SOE}}.
}
\examples{
dat <- loadDataSet("3D S Curve", n = 50)
emb <- soe@fun(dat, soe@stdpars)


plot(emb@data@data)


}
\keyword{datasets}

