% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{total_correlation,dimRedResult-method}
\alias{total_correlation,dimRedResult-method}
\alias{total_correlation}
\title{Method total_correlation}
\usage{
\S4method{total_correlation}{dimRedResult}(
  object,
  naxes = ndims(object),
  cor_method = "pearson",
  is.rotated = FALSE
)
}
\arguments{
\item{object}{of class dimRedResult}

\item{naxes}{the number of axes to use for optimization.}

\item{cor_method}{the correlation method to use.}

\item{is.rotated}{if FALSE the object is rotated.}
}
\description{
Calculate the total correlation of the variables with the axes to
assess the quality of a dimensionality reduction.
}
\seealso{
Other Quality scores for dimensionality reduction: 
\code{\link{AUC_lnK_R_NX,dimRedResult-method}},
\code{\link{LCMC,dimRedResult-method}},
\code{\link{Q_NX,dimRedResult-method}},
\code{\link{Q_global,dimRedResult-method}},
\code{\link{Q_local,dimRedResult-method}},
\code{\link{R_NX,dimRedResult-method}},
\code{\link{cophenetic_correlation,dimRedResult-method}},
\code{\link{distance_correlation,dimRedResult-method}},
\code{\link{mean_R_NX,dimRedResult-method}},
\code{\link{plot_R_NX}()},
\code{\link{quality,dimRedResult-method}},
\code{\link{reconstruction_error,dimRedResult-method}},
\code{\link{reconstruction_rmse,dimRedResult-method}}
}
\concept{Quality scores for dimensionality reduction}
