% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\docType{class}
\name{tSNE-class}
\alias{tSNE}
\alias{tSNE-class}
\title{t-Distributed Stochastic Neighborhood Embedding}
\description{
Instance of \code{\link{dimRedMethod}} for t-Distributed Stochastic Neighborhood Embedding.
}
\details{
For details see \code{\link[Rtsne]{Rtsne}}
}
\examples{
dat <- loadDataSet("3D S Curve")
tsne <- tSNE()
emb <- tsne@fun(dat, tsne@stdpars)


plot(emb@data@data)

}

