% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\docType{data}
\name{tsne}
\alias{tsne}
\title{t-Distributed Stochastic Neighborhood Embedding}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
tsne
}
\description{
Instance of \code{\link{dimRedMethod}} for t-Distributed Stochastic Neighborhood Embedding.
}
\details{
For details see \code{\link[Rtsne]{Rtsne}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- tsne@fun(dat, tsne@stdpars)


plot(emb@data@data)

}
\keyword{datasets}

