% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Converts to data.frame}
\usage{
as.data.frame(x, row.names, optional, ...)
}
\arguments{
\item{x}{The object to be converted}

\item{row.names}{unused in \code{dimRed}}

\item{optional}{unused in \code{dimRed}}

\item{...}{other arguments.}
}
\description{
General conversions of objects created by \code{dimRed} to \code{data.frame}
}

