% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedData-class.R, R/dimRedResult-class.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,dimRedData-method}
\alias{as.data.frame,dimRedResult-method}
\title{Convert to \code{data.frame}}
\usage{
as.data.frame(x, row.names, optional, ...)

\S4method{as.data.frame}{dimRedData}(x, meta.prefix = "meta.",
  data.prefix = "")

\S4method{as.data.frame}{dimRedResult}(x, org.data.prefix = "org.",
  meta.prefix = "meta.", dimRed.prefix = "")
}
\arguments{
\item{x}{dimRedResult object.}

\item{row.names}{unused}

\item{optional}{unused}

\item{...}{unused}

\item{meta.prefix}{prefix for the meta slot}

\item{data.prefix}{prefix for the dim Red slot}

\item{org.data.prefix}{prefix for the org.data slot}

\item{dimRed.prefix}{prefix for the dim Red slot}
}
\description{
Convert a dimRedData object to a data.frame.

Convert a dimRedResult object to a data.frame.
}
\details{
To avoid column name collisions prefixes for each slot can be given.

To avoid column name colissions prefixes for each slot can be given.
}
\examples{
as.data.frame(loadDataSet("Iris"), meta.prefix = "")

}
\seealso{
Other dimRedData: \code{\link{as.dimRedData}},
  \code{\link{dimRedData}}

Other dimRedResult: \code{\link{dimRedResult-class}}
}

