% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/dimRedData-class.R
\name{as.dimRedData}
\alias{as.dimRedData}
\alias{as.dimRedData,formula-method}
\title{Converts to dimRedData}
\usage{
as.dimRedData(formula, ...)

\S4method{as.dimRedData}{formula}(formula, data)
}
\arguments{
\item{formula}{The formula, left hand side is assigned to the meta slot right
hand side is assigned to the data slot.}

\item{...}{other arguments.}

\item{data}{Will be coerced into a \code{\link{data.frame}} with
\code{\link{as.data.frame}}}
}
\description{
Conversion functions to dimRedData.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: Convert a \code{data.frame} to a dimRedData
object using a formula
}}

\examples{
## create a dimRedData object using a formula
as.dimRedData(Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
              iris)[1:5]

}
\seealso{
Other dimRedData: 
\code{\link{dimRedData-class}}
}
\concept{dimRedData}
