% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.dimRedData}
\alias{as.dimRedData}
\title{Converts to dimRedData}
\usage{
as.dimRedData(formula, ...)
}
\arguments{
\item{formula}{a formula object.}

\item{...}{other arguments.}
}
\description{
Conversion functions to dimRedData.
}

