% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\docType{methods}
\name{AUC_lnK_R_NX,dimRedResult-method}
\alias{AUC_lnK_R_NX}
\alias{AUC_lnK_R_NX,dimRedResult-method}
\title{Method AUC_lnK_R_NX}
\usage{
\S4method{AUC_lnK_R_NX}{dimRedResult}(object)
}
\arguments{
\item{object}{of class dimRedResult}
}
\description{
Calculate the Area under the R_NX(ln K), used in Lee et. al. (2013).
}
\references{
Lee, J.A., Renard, E., Bernard, G., Dupont, P., Verleysen, M.,
2013. Type 1 and 2 mixtures of Kullback-Leibler divergences as cost
functions in dimensionality reduction based on similarity
preservation. Neurocomputing. 112,
92-107. doi:10.1016/j.neucom.2012.12.036
}
\seealso{
Other Quality scores for dimensionality reduction: \code{\link{LCMC,dimRedResult-method}},
  \code{\link{Q_NX,dimRedResult-method}},
  \code{\link{Q_global,dimRedResult-method}},
  \code{\link{Q_local,dimRedResult-method}},
  \code{\link{R_NX,dimRedResult-method}},
  \code{\link{cophenetic_correlation,dimRedResult-method}},
  \code{\link{distance_correlation,dimRedResult-method}},
  \code{\link{mean_R_NX,dimRedResult-method}},
  \code{\link{quality,dimRedResult-method}},
  \code{\link{reconstruction_error,dimRedResult-method}},
  \code{\link{reconstruction_rmse,dimRedResult-method}},
  \code{\link{total_correlation,dimRedResult-method}}
}

