% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmap.R
\docType{data}
\name{diffmap}
\alias{diffmap}
\title{Create Diffusion Maps}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
diffmap
}
\description{
Constructs a diffuson map
}
\details{
for details see \code{\link[diffusionMap]{diffuse}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- diffmap@fun(dat)

plot(emb@data@data)

}
\keyword{datasets}

