% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmap.R
\docType{data}
\name{diffmap}
\alias{diffmap}
\title{Diffusion Maps}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
diffmap
}
\description{
Instance of \code{\link{dimRedMethod}} to construct a diffuson map.
}
\details{
Wraps around \code{\link[diffusionMap]{diffuse}}, see there for
details.

Parameters are a list with elements
\describe{
  \item{d}{a function transforming a matrix row wise into a distance matrix or \code{dist} object, e.g.\ \code{\link[stats]{dist}}.}
  \item{ndim}{The number of dimensions}
  \item{eps}{The epsilon parameter that determines the
     diffusion weight matrix from a distance matrix d,
     \eqn{exp{-d^2/eps}}, if set to \code{"auto"} it will
     be set to the median distance to the 0.01*n nearest
     neighbor.}
  \item{t}{Time-scale parameter. The recommended value, 0,
     uses multiscale geometry.}
}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- diffmap@fun(dat, diffmap@stdpars)

plot(emb, type = "2vars")

}
\keyword{datasets}

