% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\docType{package}
\name{dimRed-package}
\alias{dimRed}
\alias{dimRed-package}
\title{A package to aggregate dimensionality reduction techniques.}
\description{
dimRed collects dimensionality reduction methods that are
implemented in R and gives them a common interface.
}
\details{
Funding provided by the Max Planck Institute for Biogeochemistry,
Jena, Empirical Inference of the Earth System Group.

TODO: add a nice description
}

