% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\docType{package}
\name{dimRed-package}
\alias{dimRed}
\alias{dimRed-package}
\title{The dimRed package}
\description{
This package simplifies dimensionality reduction in R by providing
a framework in form of S4 classes and methods. dimRed collects
dimensionality reduction methods that are implemented in R and
gives them a common interface. It also provides plotting functions
for visualization and functions for quality assessment.

Funding provided by the Max Planck Institute for Biogeochemistry,
Jena, Empirical Inference of the Earth System Group.
}

