% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedData-class.R
\docType{class}
\name{dimRedData-class}
\alias{[,dimRedData,ANY,ANY,ANY-method}
\alias{as.data.frame,dimRedData-method}
\alias{as.dimRedData,formula-method}
\alias{dimRedData}
\alias{dimRedData-class}
\alias{getData,dimRedData-method}
\alias{getMeta,dimRedData-method}
\alias{nrow,dimRedData-method}
\title{Class "dimRedData"}
\usage{
\S4method{as.data.frame}{dimRedData}(x, meta.prefix = "meta.",
  data.prefix = "")

\S4method{as.dimRedData}{formula}(formula, data)

\S4method{getData}{dimRedData}(object)

\S4method{getMeta}{dimRedData}(object)

\S4method{nrow}{dimRedData}(x)

\S4method{[}{dimRedData,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{Of class dimRedData}

\item{meta.prefix}{Prefix for the columns of the meta data names.}

\item{data.prefix}{Prefix for the columns of the variable names.}

\item{formula}{The formula, left hand side is assigned to the meta slot
right hand side is assigned to the data slot.}

\item{data}{A data frame}

\item{object}{Of class dimRedData.}

\item{i}{a valid index.}
}
\description{
A class and methods to hold data for dimensionality reduction.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.data.frame}: convert to data.frame

\item \code{as.dimRedData}: Convert a \code{data.frame} to a dimRedData
object using a formula

\item \code{getData}: Get the data slot.

\item \code{getMeta}: Get the meta slot.

\item \code{nrow}: Get the number of observations.

\item \code{[}: Subset rows.
}}
\section{Slots}{

\describe{
\item{\code{data}}{of class \code{matrix}, holds the data, observations in
rows, variables in columns}

\item{\code{meta}}{of class \code{data.frame}, holds meta data such as
classes, internal manifold coordinates, or simply additional
data of the data set. Must have the same number of rows as the
\code{data} slot or be an empty data frame.}
}}
\examples{
s3d <- loadDataSet("3D S Curve")
as(s3d, "data.frame")
iris <- as(iris[,1:4], "dimRedData")

getData(s3d)
getMeta(s3d)
nrow(s3d)
s3d[1:40,]

as.dimRedData(Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
              iris)

}

