% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedData-class.R
\docType{class}
\name{dimRedData-class}
\alias{[,dimRedData,ANY,ANY,ANY-method}
\alias{dimRedData}
\alias{dimRedData-class}
\alias{getData,dimRedData-method}
\alias{getMeta,dimRedData-method}
\alias{nrow,dimRedData-method}
\title{An S4 class and methods to hold data for dimensionality reduction}
\usage{
\S4method{getData}{dimRedData}(object)

\S4method{getMeta}{dimRedData}(object)

\S4method{nrow}{dimRedData}(x)

\S4method{[}{dimRedData,ANY,ANY,ANY}(x, i)
}
\description{
An S4 class and methods to hold data for dimensionality reduction
}
\section{Methods (by generic)}{
\itemize{
\item \code{getData}: extract the data slot (a matrix).

\item \code{getMeta}: extract the meta slot (a data.frame).

\item \code{nrow}: get the number of data entries.

\item \code{[}: subsetting as in a matrix.
}}
\section{Slots}{

\describe{
\item{\code{data}}{of class \code{matrix}, holds the data, observations in
rows, variables in columns}

\item{\code{meta}}{of class \code{data.frame}, holds meta data such as
classes, internal manifold coordinates, or simply additional
data of the data set. Must have the same number of rows as the
\code{data} slot or be an empty data frame.}
}}
\examples{
s3d <- loadDataSet("3D S Curve")
as(s3d, "data.frame")
iris <- as(iris[,1:4], "dimRedData")

getData(s3d)
getMeta(s3d)
nrow(s3d)
s3d[1:40,]

}

