% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedMethod-class.R
\name{dimRedMethodList}
\alias{dimRedMethodList}
\title{dimRedMethodList}
\usage{
dimRedMethodList()
}
\value{
a character vector with the names of classes that inherit
    from \code{dimRedMethod}.
}
\description{
Get the names of all methods for dimensionality reduction.
}
\details{
Returns the name of all classes that inherit from
\code{\link{dimRedMethod-class}} to use with \code{\link{embed}}.
}
\examples{
dimRedMethodList()

}
