% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedResult-class.R
\docType{class}
\name{dimRedResult-class}
\alias{as.data.frame,dimRedResult-method}
\alias{dimRedResult}
\alias{dimRedResult-class}
\alias{getPars,dimRedResult-method}
\alias{print,dimRedResult-method}
\title{Class "dimRedResult"}
\usage{
\S4method{as.data.frame}{dimRedResult}(x, org.data.prefix = "org.",
  meta.prefix = "meta.", data.prefix = "")

\S4method{getPars}{dimRedResult}(object)

\S4method{print}{dimRedResult}(x)
}
\arguments{
\item{x}{Of class \code{dimRedResult}}

\item{org.data.prefix}{Prefix for the columns of the org.data slot.}

\item{meta.prefix}{Prefix for the columns of \code{x@data@meta}.}

\item{data.prefix}{Prefix for the columns of \code{x@data@data}.}

\item{object}{Of class \code{dimRedResult}}
}
\description{
A class to hold the results of of a dimensionality reduction.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.data.frame}: convert to \code{data.frame}

\item \code{getPars}: Get the parameters with which the method
was called.

\item \code{print}: Method for printing.
}}
\section{Slots}{

\describe{
\item{\code{data}}{Output data of class dimRedData.}

\item{\code{org.data}}{original data, a matrix.}

\item{\code{apply}}{a function to apply the method to out-of-sampledata,
may not exist.}

\item{\code{inverse}}{a function to calculate the original coordinates from
reduced space, may not exist.}

\item{\code{has.org.data}}{logical, if the original data is included in the object.}

\item{\code{has.apply}}{logical, if a forward method is exists.}

\item{\code{has.inverse}}{logical if an inverse method exists.}

\item{\code{method}}{saves the method used.}

\item{\code{pars}}{saves the parameters used.}
}}

