% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRedResult-class.R
\docType{class}
\name{dimRedResult-class}
\alias{dimRedResult}
\alias{dimRedResult-class}
\title{A class to hold the results of of a dimensionality reduction.}
\description{
A class to hold the results of of a dimensionality reduction.
}
\section{Slots}{

\describe{
\item{\code{data}}{Output data of class dimRedData.}

\item{\code{org.data}}{original data, a matrix.}

\item{\code{apply}}{a function to apply the method to out-of-sampledata,
may not exist.}

\item{\code{inverse}}{a function to calculate the original coordinates from
reduced space, may not exist.}

\item{\code{has.org.data}}{logical, if the original data is included in the object.}

\item{\code{has.apply}}{logical, if a forward method is exists.}

\item{\code{has.inverse}}{logical if an inverse method exists.}
}}

