% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drr.R
\docType{data}
\name{drr}
\alias{drr}
\title{Dimensionality Reduction via Regression}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
drr
}
\description{
Instance of \code{\link{dimRedMethod}} for Dimensionality Reduction via Regression.
}
\details{
For details see \code{\link[DRR]{DRR}}.
}
\examples{
dat <- loadDataSet("3D S Curve", n = 500)
emb <- drr@fun(dat, pars = list(ndim = 2))

plot(emb@data@data)


}
\keyword{datasets}

