% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomap.R
\docType{data}
\name{isomap}
\alias{isomap}
\title{Isomap embedding}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
isomap
}
\description{
Fit the isomap embedding.
}
\details{
For details see \code{\link[vegan]{isomap}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- isomap@fun(dat)

plot(emb@data@data)

}
\keyword{datasets}

