% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embed.R
\docType{data}
\name{kamada_kawai}
\alias{kamada_kawai}
\title{Graph Embedding via the Kamada Kawai Algorithm}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
kamada_kawai
}
\description{
Instance of \code{\link{dimRedMethod}} for the Kamada Kawai Algorithm.
}
\details{
For details see \code{\link[igraph]{layout_with_kk}}.
}
\examples{
dat <- loadDataSet("Swiss Roll")
kk <- kamada_kawai@fun(dat, kamada_kawai@stdpars)

plot(kk@data@data)


}
\keyword{datasets}

