% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\docType{data}
\name{kpca}
\alias{kpca}
\title{implements Kernel PCA for the use with dimRed}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
kpca
}
\arguments{
\item{data}{an object of class \code{\link{dimRedData-class}}.}

\item{pars}{an object of class \code{\link{dimRedMethodPars-class}}}
}
\value{
an object of class \code{\link{dimRedResult-class}}
}
\description{
uses \code{\link[kernlab]{kpca}} internally. Has methods for
forward and inverse calculations.
}
\examples{

dat <- loadDataSet("3D S Curve")
emb <- kpca@fun(dat, list(kernel = 'rbfdot',
                          kpar = list(sigma = 0.1),
                          ndim = 2))

plot(emb@data@data)

}
\keyword{datasets}

