% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leim.R
\docType{data}
\name{leim}
\alias{leim}
\title{Laplacian Eigenmaps}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
leim
}
\description{
Instance of \code{\link{dimRedMethod}} for Laplacian Eigenmaps.
}
\details{
For details see \code{\link[loe]{spec.emb}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- leim@fun(dat, leim@stdpars)


plot(emb@data@data)


}
\keyword{datasets}

