% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lle.R
\docType{data}
\name{lle}
\alias{lle}
\title{Local linear embedding}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
lle
}
\description{
Fit a local linear embedding
}
\details{
for details see \code{\link[lle]{lle}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- lle@fun(dat)


plot(emb@data@data)

}
\keyword{datasets}

