% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\docType{data}
\name{mds}
\alias{mds}
\title{Metric dimensional scaling}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
mds
}
\description{
Fit a metric dimensional scaling
}
\details{
for details see \code{\link[stats]{cmdscale}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- mds@fun(dat)

plot(emb@data@data)

}
\keyword{datasets}

