% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\docType{data}
\name{nmds}
\alias{nmds}
\title{Non-metric dimensional scaling}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
nmds
}
\description{
Implements non metric dimensional scaling
}
\details{
for details see \code{\link[vegan]{monoMDS}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- nmds@fun(dat)

plot(emb@data@data)

}
\keyword{datasets}

