% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\docType{data}
\name{pca}
\alias{pca}
\title{implements PCA for the use with dimRed}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
pca
}
\arguments{
\item{data}{an object of class \code{\link{dimRedData-class}}.}

\item{pars}{a list with parameters}
}
\value{
an object of class \code{\link{dimRedResult-class}}
}
\description{
uses the \code{\link{prcomp}} function internally. Has methods for
forward and inverse calculations.
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- pca@fun(dat, pca@stdpars)

plot(emb, type = "2vars")
plot(emb@inverse(emb@data), type = "3vars")

}
\keyword{datasets}

