% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\docType{data}
\name{pca}
\alias{pca}
\title{Principal Component Analysis}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
pca
}
\description{
Instance of \code{\link{dimRedMethod}} for Principal Component Analysis.
}
\details{
For details see \code{\link{prcomp}}.
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- pca@fun(dat, pca@stdpars)

plot(emb, type = "2vars")
plot(emb@inverse(emb@data), type = "3vars")

}
\keyword{datasets}

