% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\docType{methods}
\name{quality}
\alias{Q_global}
\alias{Q_global,dimRedResult-method}
\alias{Q_local}
\alias{Q_local,dimRedResult-method}
\alias{cophenetic_correlation}
\alias{cophenetic_correlation,dimRedResult,missingORfunction,missingORcharacter-method}
\alias{dimRedQualityList}
\alias{distance_correlation}
\alias{distance_correlation,dimRedResult-method}
\alias{mean_R_NX}
\alias{mean_R_NX,dimRedResult-method}
\alias{quality}
\alias{reconstruction_rmse}
\alias{reconstruction_rmse,dimRedResult-method}
\alias{total_correlation}
\alias{total_correlation,dimRedResult,missingORnumeric,missingORcharacter,missingORlogical-method}
\title{dispatches the different methods for quality assessment}
\usage{
quality(.data, .method = dimRedQualityList(), .mute = character(0), ...)

\S4method{Q_local}{dimRedResult}(object)

\S4method{Q_global}{dimRedResult}(object)

\S4method{mean_R_NX}{dimRedResult}(object)


  \S4method{total_correlation}{dimRedResult,missingORnumeric,missingORcharacter,missingORlogical}(object,
  naxes, cor_method, is.rotated)


  \S4method{cophenetic_correlation}{dimRedResult,missingORfunction,missingORcharacter}(object,
  d, cor_method)

\S4method{distance_correlation}{dimRedResult}(object)

\S4method{reconstruction_rmse}{dimRedResult}(object)

dimRedQualityList()
}
\arguments{
\item{.data}{object of class \code{dimRedResult}}

\item{.method}{character vector naming one of the methods}

\item{.mute}{what output from the embedding method should be muted.}

\item{...}{the pameters, internally passed as a list to the
quality method as \code{pars = list(...)}}

\item{object}{of class dimRedResult}

\item{naxes}{the number of axes to use for optimization.}

\item{cor_method}{the correlation method to use.}

\item{is.rotated}{if FALSE the object is rotated.}

\item{d}{the distance function to use.}
}
\value{
a number
}
\description{
wraps around all quality assessment functions functions.
}
\details{
method must be one of \code{"\link{Q_local}", "\link{Q_global}",
"\link{mean_R_NX}", "\link{total_correlation}",
"\link{cophenetic_correlation}", "\link{distance_correlation}",
"\link{reconstruction_rmse}"}
}
\examples{
embed_methods <- dimRedMethodList() 
quality_methods <- dimRedQualityList()
scurve <- loadDataSet("3D S Curve", n = 500)

quality_results <- matrix(NA, length(embed_methods), length(quality_methods),
                              dimnames = list(embed_methods, quality_methods))
embedded_data <- list()

for(e in embed_methods) {
  message("embedding: ", e)
  embedded_data[[e]] <- embed(scurve, e, .mute = c("message", "output"))
  for(q in quality_methods) {
    message("  quality: ", q)
    quality_results[e,q] <- tryCatch(
      quality(embedded_data[[e]], q), 
      error = function (e) NA
    )
  }
}

print(quality_results)

}

