% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soe.R
\docType{data}
\name{soe}
\alias{soe}
\title{Soft ordinal embedding}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
soe
}
\description{
implements soft ordinal embedding
}
\details{
for details see \code{\link[loe]{SOE}}
}
\examples{
dat <- loadDataSet("3D S Curve", n = 50)
emb <- soe@fun(dat)


plot(emb@data@data)


}
\keyword{datasets}

