% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\docType{data}
\name{tsne}
\alias{tsne}
\title{t distributed stochastic neighborhood embedding}
\format{An object of class \code{dimRedMethod} of length 1.}
\usage{
tsne
}
\description{
implements t-SNE
}
\details{
for details see \code{\link[Rtsne]{Rtsne}}
}
\examples{
dat <- loadDataSet("3D S Curve")
emb <- tsne@fun(dat)


plot(emb@data@data)



}
\keyword{datasets}

